Separate LoadingIndicator logic from the DisplayTarget
