+++ android/activity/src/main/java/com/jrmitchell/cube/investigation/jitpack/modular/test/android/activity/DefaultActivityDisplayTarget.kt
@@ -5,13 +5,15 @@ import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import com.jrmitchell.cube.investigation.jitpack.modular.test.android.core.abstract.AndroidDisplayTarget
 import com.jrmitchell.cube.investigation.jitpack.modular.test.android.core.abstract.AndroidImageLoader
+import com.jrmitchell.cube.investigation.jitpack.modular.test.android.core.abstract.AndroidLoadingIndicator
 import com.jrmitchell.cube.investigation.jitpack.modular.test.android.core.resolver.MatchActionResolver
 import com.jrmitchell.cube.investigation.jitpack.modular.test.android.core.resolver.MultiActionResolver
+import com.jrmitchell.cube.investigation.jitpack.modular.test.core.data.DisplayData
 
 /**
  * [AppCompatActivity] and [AndroidDisplayTarget] implementation based on some assumptions of how actions will be resolved and errors shown
  */
-abstract class DefaultActivityDisplayTarget <L : AndroidImageLoader> : AppCompatActivity(), AndroidDisplayTarget<MultiActionResolver, L> {
+abstract class DefaultActivityDisplayTarget <L : AndroidImageLoader> : AppCompatActivity(), AndroidDisplayTarget<MultiActionResolver, L>, AndroidLoadingIndicator {
 	
 	companion object {
 		const val ACTION_ID_EXTRA_KEY = "ACTION_ID_EXTRA_KEY"
@@ -27,8 +29,13 @@ abstract class DefaultActivityDisplayTarget <L : AndroidImageLoader> : AppCompat
 		resolvers.add(MatchActionResolver("""close""".toRegex()) { _, _ -> finishAffinity() })
 	}
 	
+	override fun displayData(data: DisplayData) {
+		setLoadingState(false)
+		super.displayData(data)
+	}
+	
 	override fun displayError(throwable: Throwable?) {
-		super.displayError(throwable)
+		setLoadingState(false)
 		Toast.makeText(this, throwable?.localizedMessage ?: "Unknown error", Toast.LENGTH_LONG).show()
 	}
