+++ android/core/src/main/java/com/jrmitchell/cube/investigation/jitpack/modular/test/android/core/abstract/AndroidDisplayTarget.kt
@@ -1,9 +1,7 @@
 package com.jrmitchell.cube.investigation.jitpack.modular.test.android.core.abstract
 
-import android.view.View
 import android.widget.ImageView
 import android.widget.TextView
-import androidx.core.view.isVisible
 import androidx.recyclerview.widget.RecyclerView
 import com.jrmitchell.cube.investigation.jitpack.modular.test.android.core.adapter.ButtonAdapter
 import com.jrmitchell.cube.investigation.jitpack.modular.test.core.abstract.DisplayTarget
@@ -13,11 +11,6 @@ import com.jrmitchell.cube.investigation.jitpack.modular.test.core.data.DisplayD
  * Interface for an Android target where data can be displayed
  */
 interface AndroidDisplayTarget<R : AndroidActionResolver, L : AndroidImageLoader> : DisplayTarget {
-	/**
-	 * View to be used as loading UI
-	 */
-	val loadingUi: View
-	
 	/**
 	 * View to display the data's main image in
 	 */
@@ -49,19 +42,10 @@ interface AndroidDisplayTarget<R : AndroidActionResolver, L : AndroidImageLoader
 	val imageLoader: L
 	
 	override fun displayData(data: DisplayData) {
-		setLoadingState(false)
 		titleView.text = data.titleText
 		buttonAdapter.data = data.buttonData
 		imageView.contentDescription = data.imageContentDescription
 		imageLoader.loadImage(data.imageData, imageView)
-	}
-	
-	override fun displayError(throwable: Throwable?) {
-		setLoadingState(false)
-	}
-	
-	override fun setLoadingState(isLoading: Boolean) {
-		loadingUi.isVisible = isLoading
 	}
 	
 	/**
