+++ core/src/main/java/com/jrmitchell/cube/investigation/jitpack/modular/test/core/abstract/LoadingIndicator.kt
@@ -1,0 +1,11 @@
+package com.jrmitchell.cube.investigation.jitpack.modular.test.core.abstract
+
+/**
+ * Interface for a generic indicator of loading state for data display
+ */
+interface LoadingIndicator {
+	/**
+	 * Set the loading state for displayed data
+	 */
+	fun setLoadingState(isLoading : Boolean)
+}
